ch06

+++ Textbook_PART1/src/ch03/sec01/SignOperatorExample.java
@@ -7,7 +7,7 @@ package ch03.sec01;
 public class SignOperatorExample {
 
 	public static void main(String[] args) {
-		int x = -100;
+	byte x = -100;
 		x = -x;
 		System.out.println("x : " +x);
